fix:디테일 페이지 수정 협의 피요
